refactor(systems): 環境ログを turn_system.go に統合する
logEnvironmentChange は runEndPhase からしか呼ばれず、ターン終了処理の
一部なので独立ファイルにする必要がない。呼び出し元と同じファイルへ寄せて
流れを1箇所で追えるようにする。テストも turn_system_test.go へ移す。
