+++ docs/design/260830162748.md
@@ -34,7 +34,7 @@ func (gt *GameTime) TimeOfDayJustChanged() (TimeOfDay, bool)
 
 いずれも内部で `prev := GameTime{TotalTurns: gt.TotalTurns - 1}` を作り、既存の `GetSeason` / `GetTimeOfDay` に委譲する。導出の計算式を複製しない。
 
-ログ出力は `internal/systems` の関数 `logEnvironmentChange(world)` に置き、ターン確定時に1度だけ呼ぶ。日の出入りの割り当ては presentation の方針なので、`GameTime` でなくこの関数が持つ。夜明けへ入るのが日の出、夕へ入るのが日の入りとし、ほかの区分への変化は出さない。
+ログ出力は `turn_system.go` の関数 `logEnvironmentChange(world)` に置き、ターン確定時に1度だけ呼ぶ。呼び出し元の `runEndPhase` と同じファイルにまとめ、ターン終了処理の流れを1箇所で追えるようにする。日の出入りの割り当ては presentation の方針なので、`GameTime` でなくこの関数が持つ。夜明けへ入るのが日の出、夕へ入るのが日の入りとし、ほかの区分への変化は出さない。
 
 ```go
 if tod, changed := gt.TimeOfDayJustChanged(); changed {
@@ -55,11 +55,10 @@ if tod, changed := gt.TimeOfDayJustChanged(); changed {
 | ファイル | 変更内容 |
 |---|---|
 | `internal/components/game_time.go` | `SeasonJustChanged` と `TimeOfDayJustChanged` を追加 |
-| `internal/systems/environment_log.go` | 新規。`logEnvironmentChange` を定義 |
-| `internal/systems/turn_system.go` | `runEndPhase` の `Advance` 直後に `logEnvironmentChange` を呼ぶ |
+| `internal/systems/turn_system.go` | `logEnvironmentChange` を定義し、`runEndPhase` の `Advance` 直後に呼ぶ |
 | `internal/i18n/locale/ja.po` | 3つの msgid を追加。季節名と時間帯名は既存 |
 | `internal/components/game_time_test.go` | 2つの判定メソッドの table テストを追加 |
-| `internal/systems/environment_log_test.go` | 新規。境界ターンでのログ出力を検証 |
+| `internal/systems/turn_system_test.go` | 境界ターンでのログ出力を検証する `TestLogEnvironmentChange` を追加 |
 
 ## 採用しなかった方法
